moved to modules
-all of the previosly made folders have been moved to a modules folder, for a clear distinction of modules from output folders etc.
