+++ CMakeLists.txt
@@ -51,20 +51,25 @@ add_executable (
     ${PROJECT_NAME} 
         "main.cpp" 
         "platinumEyes.h" 
-        "utilities/functions.cpp" 
-        "utilities/trim.cpp" 
-        "gui/guiUtilities.cpp" 
-        "keybindings/keyBindings.cpp" 
-        "config/configUtilities.cpp" 
-        "ocr/readingItems.cpp" 
-        "gui/settingsMenu.cpp"
-        "ocr/tesseract.cpp"
-        "ocr/easyocr.cpp" 
-        "ocr/ocrfunctions.cpp" 
-        "gui/settingsChanges.cpp" 
-        "gui/settingsChanged.h" 
-        "relics/relicDatabase.cpp" 
-         "ocr/ocr.h" "config/config.h" "utilities/utilities.h" "gui/gui.h" "keybindings/keybindings.h" "relics/relics.h")
+        "modules/utilities/functions.cpp" 
+        "modules/utilities/trim.cpp" 
+        "modules/gui/guiUtilities.cpp" 
+        "modules/keybindings/keyBindings.cpp" 
+        "modules/config/configUtilities.cpp" 
+        "modules/ocr/readingItems.cpp" 
+        "modules/gui/settingsMenu.cpp"
+        "modules/ocr/tesseract.cpp"
+        "modules/ocr/easyocr.cpp" 
+        "modules/ocr/ocrfunctions.cpp" 
+        "modules/gui/settingsChanges.cpp" 
+        "modules/gui/settingsChanged.h" 
+        "modules/relics/relicDatabase.cpp" 
+        "modules/ocr/ocr.h" 
+        "modules/config/config.h" 
+        "modules/utilities/utilities.h" 
+        "modules/gui/gui.h" 
+        "modules/keybindings/keybindings.h" 
+        "modules/relics/relics.h")
 
 
 target_link_libraries(${PROJECT_NAME} PRIVATE ${LIBRARIES})
