+++ modules/ocr/easyocr.cpp
