+++ modules/ocr/ocr.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "../platinumEyes.h"
+#include "../../platinumEyes.h"
 
 
 BOOL SaveHBITMAPToFile(HBITMAP hBitmap, LPCTSTR lpszFileName);
