+++ modules/ocr/ocrfunctions.cpp
