+++ modules/ocr/readingItems.cpp
