+++ modules/ocr/tesseract.cpp
