+++ modules/relics/relics.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "../platinumEyes.h"
+#include "../../platinumEyes.h"
 
 
 int parseRelicData();
